add cpl a and stop $xx
As in RGBASM, ca83 ALU instructions tolerate presence or absence
of explicit a, argument.  WLA DX accepts both as well since 2019-12
(vhelin/wla-dx#271).  For example,  these assemble the same:

    add 69
    add a, 69

On 2020-05-05, in the gbdev Discord server, user Lifeless_G
complained of an RGBASM syntax error from this instruction:

    cpl a

So I made it do the expected thing: same as cpl.

Others have mentioned use of a signature byte after a stop
instruction.  The CPU ignores the following byte, but a tool
may put information there.  So I gave stop a signature argument,
much as brk on 6502 family has:

    stop $00
    stop      ; same as stop $00
    stop $DB

Finally, I found and fixed serious problems in the makefile.