reject VBA for pak swap breakage
Border Crossing uses di halt, a pair of instructions that wait
for an interrupt without calling the interrupt handler.  This is
because after the cartridge is unplugged, the interrupt handler
is in another castle.  VBA implements di halt wrong, calling the
interrupt handler anyway, causing incorrect flashing of the
throbber dot in the pak swap screen.  Detect this misbehavior
and show an error message.