Border Crossing: reduce fadeout flicker at bottom
On 2023-03-23, dalton in the SNESdev server reported that the SGB
border's bottom scanline was flickering while it processed some
packets.  I investigated in Mesen and found an off-by-one in SGB's
background scroll value causing the top scanline of the tilemap
row below the bottom of the border to become visible.  To make it
less visible, duplicate the border's bottom row (V flipped).