Border Crossing: wait for vblank before swapping
The wait loop first draws the blinking dot in the corrner.
This was causing intermittent writes to inaccessible VRAM when
the status screen first appeared after loading a border.
This was harmless on hardware, as the dot would get drawn
correctly a quarter second later, but bgb was raising exceptions.
Waiting for vertical blanking before jumping into the wait loop
fixed VRAM inaccessibility.

Also add `def` before `equ` for compatibility with RGBDS 0.7+